Add Type Summary for StyleSelfAlignmentData. https://bugs.webkit.org/show_bug.cgi?id=295650 rdar://problem/155454737
Reviewed by Ryan Reno.

StyleSelfAlignmentData has bitfields used to represent ItemPosition,
ItemPositionType, and OverflowAlignment which are all enums. Printing
this class in lldb right now is not super helpful and you would need to
call the respective functions, which return the enum, to get any useful
information easily. We can add a type summary for this that does the
same thing so that it makes debugging a bit easier.

Canonical link: https://commits.webkit.org/297184@main