Introduce SupplementBase as a way to downcast Supplement<T> https://bugs.webkit.org/show_bug.cgi?id=295721 rdar://155542045
Reviewed by Chris Dumez.

For the sake of safety, we often prefer to use downcast in place of
static_cast (the static analyzer catches cases where we should).

In order to use downcast, we may require a TypeCastsTraits specialization
which has an isType() function. This function takes in an argument with
the type of the base class and calls a function that returns true if the
instance is of the derived class and false if it's not.

Consider the case where class Bar inherits from Supplement<Foo> and we
want to downcast from Supplement<Foo> to Bar. Normally, Supplement<Foo>
would have a virtual function isBar() that returns false, and Bar would
override this function to return true. The isType() function in the
TypeCastsTraits specialization would call isBar() to determine if an
instance of Supplement<Foo> is also an instance of Bar.

But Supplement<Foo> is a class generated by the Supplement<T> template.
We cannot declare and implement the virtual function isBar() in the
template because class templates cannot contain virtual functions.

We fix this by creating a concrete base class SupplementBase that the
template Supplement<T> inherits from. Virtual functions like isBar() can be
declared + implemented here.

This patch adds this class and uses it to fix the static analysis warning
in NavigatorCookieConsent.cpp.

(Thank you to Wenson for helping come up with the idea).

* Source/WebCore/Modules/cookie-consent/NavigatorCookieConsent.cpp:
(WebCore::NavigatorCookieConsent::from):
* Source/WebCore/Modules/cookie-consent/NavigatorCookieConsent.h:
(isType):
* Source/WebCore/SaferCPPExpectations/MemoryUnsafeCastCheckerExpectations:
* Source/WebCore/platform/Supplementable.h:
(WebCore::SupplementBase::isNavigatorCookieConsent const):

Canonical link: https://commits.webkit.org/297255@main