[Skia] Cleanup GraphicsContextSkia https://bugs.webkit.org/show_bug.cgi?id=295702
Reviewed by Carlos Garcia Campos.

This change removes invalid fixme and makes the didUpdate*State()
functions noop as in skia the change flags of state are not used.

The fixme turned out to be invalid as in skia's case - contrary to
cairo - we're building skia-level painting state (SkPaint) in a lazy
fashion always.

Canonical link: https://commits.webkit.org/297258@main