[WPE] Click count leaking through successive layout test within the same WPEView https://bugs.webkit.org/show_bug.cgi?id=295252
Reviewed by Adrian Perez de Castro.

Tests are expected to start with reset values for the click counter.
This is true for other ports that compute the click count in
EventSenderProxy because a new instance is created for every test even
in the same WebKitTestRunner. We use WPEView to compute the click
count, that is reused for all the tests in the same WebKitTestRunner.
This patch uses the same logic to compute the click count from
EventSenderProxy as all other ports do.

Canonical link: https://commits.webkit.org/297259@main