Decouple absolute positioned box auto margin distribution from alignment https://bugs.webkit.org/show_bug.cgi?id=295753 rdar://155572208
Reviewed by Alan Baradlay.

PositionedLayoutConstraints::resolvePosition attempts to resolve the
position of the positioned box. This is done in two main ways:

1.) Distributing the extra space to any auto margins that the box has in
the axis.
2.) Aligning the margin box according to its self-alignment.

Currently, the logic for these two steps is somewhat mixed together. In
particular, if the box has both of its insets fixed, then we will
attempt to see if there are automatic margins to distribute space to and
then attempt to align the box if that fails. Instead, let's factor this
out so that we can have two distinct sections of code for each of the
steps described above. This is in preparation for building upon the
alignment portion of the code incrementally.

* Source/WebCore/rendering/PositionedLayoutConstraints.cpp:
(WebCore::PositionedLayoutConstraints::resolvePosition const):

Canonical link: https://commits.webkit.org/297267@main