[Linux] Use -fno-omit-frame-pointer on developer builds. https://bugs.webkit.org/show_bug.cgi?id=295788
Reviewed by Michael Catanzaro.

This is helpful alongside the new useGdbJITInfo option for local debugging
of jit code. The JIT Dwarf info doesn't give a way to reconstruct SP, so
we need frame pointers for it to work well.

We don't enable frame pointers for all builds yet, since it might have
performance implications. We should try to do that in the future.

This was rolled out (295507@main) last time because I also included ALT_ENTRY.
This is the smallest possible patch, so hopefuly it won't break anything.

Canonical link: https://commits.webkit.org/297281@main