Anchor-positioned, fixed-positioned element might get clipped by the viewport rdar://154908754 https://bugs.webkit.org/show_bug.cgi?id=295322
Reviewed by Simon Fraser.

RenderLayerBacking::shouldClipCompositedBounds() determines whether to clip
the RenderLayer or not (in the case of a fixed-positioned element, clip to the
viewport). It calls RenderLayer::isTransformed() to check if the RenderLayer
have transforms, and it wouldn't clip if there're transforms
(RenderLayerBacking::shouldClipCompositedBounds() ->
layerOrAncestorIsTransformedOrUsingCompositedScrolling ->
RenderLayer::isTransformed())

A RenderLayer has transforms if the renderer has transforms, but also when
"scroll compensation" [1] is applied. This is part of anchor positioning,
to adjust the position of anchor-positioned boxes in response to its
anchor scrolling. The adjustment is applied as a transform, so a RenderLayer
where the renderer doesn't have transforms could still have transforms.
But RenderLayer::isTransformed() wasn't updated to include this new condition.
The result: if an anchor-positioned, fixed-positioned element got scroll
compensated, it'll be clipped by the viewport (it shouldn't)

[1]: https://drafts.csswg.org/css-anchor-position-1/#compensate-for-scroll

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-fixedpos-003-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-fixedpos-003.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/reference/anchor-scroll-fixedpos-003-ref.html: Added.

* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateTransform):
    - Use RenderLayer::isTransformed() to get whether the layer has transforms,
      instead of repeating the code.

* Source/WebCore/rendering/RenderLayer.h:
    - Added comments about scroll compensation behavior.

* Source/WebCore/rendering/RenderLayerInlines.h:
(WebCore::RenderLayer::isTransformed const):
    - RenderLayer has transforms if there's scroll compensation too.

Canonical link: https://commits.webkit.org/297297@main