[macOS] [DumpRenderTree] Several editing/pasteboard/entries-api tests are flaky timeouts https://bugs.webkit.org/show_bug.cgi?id=295816
Reviewed by Abrar Rahman Protyasha.

This is a speculative fix to try and stabilize several flaky WebKit1 layout tests — some of these
tests occasionally time out, because a drop is never handled when dispatching `mouseUp`, after
simulating a drag session.

This happens when `-[NSEvent locationInWindow]` for the synthesized mouse event returns an offset
that is far offscreen — e.g. `(-9804, 10726)` — instead of the real offset relative to the window.
After a bit more investigation, the concatenated window transform (`CGSGetCatenatedWindowTransform`)
on `DumpRenderTreeWindow` ends up having an offset close to the origin (0, 0) in the failing case,
rather than large positive x and y translation offsets in the passing case. This is important
because the value of `-locationInWindow` is internally round-tripped in AppKit through
`-[NSEvent _initWithCGEvent:eventRef:]`, which means we end up taking `newMousePosition`, mapping
it to screen coordinates when extracting the `CGEventRef`, and then have AppKit try to map the point
in screen coordinates back to window coordinates in that internal initializer.

Instead of going through all the above, we can instead simply call `-hitTest:` using our original
mouse position which is already in window coordinates.

* Tools/DumpRenderTree/mac/EventSendingController.mm:
(-[EventSendingController mouseMoveToX:Y:]):

Canonical link: https://commits.webkit.org/297298@main