[SwiftUI] An Objective-C exception may sometimes be thrown when using a URLSchemeHandler that is synchronous https://bugs.webkit.org/show_bug.cgi?id=295829 rdar://155647966
Reviewed by Wenson Hsieh.

297220@main fixed the version of this bug that happens when using Swift concurrency in a URLSchemeHandler by
using asynchronous functions and Tasks. This fix worked in this case because cancelling the scheme task resulted
in the handler's continuation terminating and skipping the for-try-await loop.

However, in the case of a purely synchronous implementation of URLSchemeHandler, the issue may still happen if
there is any expensive computation. This happens because unlike the async version, doing sync work is blocking,
and so the continuation doesn't actually end up terminating, and therefore the loop body gets executed.

Fix by checking for Task cancellation inside the loop in addition to after it.

Tested by updating an existing test to test both sync and async versions.

* Source/WebKit/UIProcess/Cocoa/WKURLSchemeHandlerAdapter.swift:
(WKURLSchemeHandlerAdapter.webView(_:start:)):
* Tools/TestWebKitAPI/Tests/WebKit Swift/URLSchemeHandlerTests.swift:
(reply(for:)):

Canonical link: https://commits.webkit.org/297301@main