AX: WebKit should match other browsers, and the main-thread text implementation, and not generate text positions for SVG graphics elements https://bugs.webkit.org/show_bug.cgi?id=295802 rdar://155637541
Reviewed by Joshua Hoffman.

This discrepancy was discovered when investigating an infinite word-navigation loop with VoiceOver in which we would
constantly cycle between the last two words. This fixes that issue, but is likely not the root cause — invesigation
will continue beyond this commit. But this behavior change is right regardless, as it matches other browsers and the
WebKit main-thread text implementation.

Minimal reproduction:

```
<div style="width: 580px">
<svg>
    <path d="M0 0" fill="#252526"/>
    <path d="M1000 1000" fill="#252526"/>
    <path d="M50 50" fill="#252526"/>
    <path d="M10 10" fill="#252526"/>
    <path d="M42 42" fill="#252526"/>
    <path d="M9999 9999" fill="#252526"/>
</svg>
<p>The quick fox</p>
</div>
```

* LayoutTests/accessibility/mac/index-for-text-markers-svg-graphics-elements-expected.txt: Added.
* LayoutTests/accessibility/mac/index-for-text-markers-svg-graphics-elements.html: Added.
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::textRuns):

Canonical link: https://commits.webkit.org/297307@main