Unnecessary calls to hasAvailableFrameCallback https://bugs.webkit.org/show_bug.cgi?id=295704 rdar://155519872
Reviewed by Youenn Fablet.

hasAvailableFrameCallback was called whenever we render a frame on screen with no exception.

This callback is used for two purposes:
1- Let the MediaPlayerPrivate knows that the first frame has been decoded and rendered
2- When gathering frame metadata due to a call to requestVideoFrameCallback

1- is only required once after a flush
2- is only required if the JS has used requestVideoFrameCallback API.

We optimise the code to only call hasAvailableFrameCallback when needed only
by creating two callbacks that can be set, the first to be notified when the
first frame has been queued for rendering, and the second to be notified
after every single frame being rendered.

No change in observable behaviour, covered by existing tests.

* LayoutTests/media/media-source/media-source-rvfc-paused-offscreen.html: Fix incorrect include path.
* LayoutTests/media/media-source/media-source-rvfc-paused.html: ditto
* LayoutTests/media/media-source/media-source-rvfc-playing-offscreen.html: ditto
* LayoutTests/media/media-source/media-source-rvfc-playing.html: ditto

Canonical link: https://commits.webkit.org/297330@main