[JSC] Fix Uint8Array fromBase64 and setFromBase64 for invalid 1-chunk input https://bugs.webkit.org/show_bug.cgi?id=295578 rdar://155346981
Reviewed by Yusuke Suzuki.

This PR updates simdutf to 7.3.3, fixes handling of invalid 1-chunk input in
base64 decoding, and removes the slow fallback.

First, base64 input is changed to unconditionally decode even when the expected
decoded binary size is 0. Previously, there was a fast path in fromBase64 input
to an empty Uint8Array when the expected decoded binary size is 0. This is
incorrect when the input contains a single invalid chunk, as it skips decoding
entirely when it should attempt to and report the error.

Second, use simdutf to decode base64 for all values of lastChunkHandling.
Previously, simdutf was only used for 'loose', and a slow fallback was used for
'strict' and 'stop-before-partial'. Now simdutf has support for all 3, so use
it and remove the fallback.

Note that simdutf currently has a bug for the read count when lastChunkHandling
is 'stop-before-partial', which this patch works around.

Canonical link: https://commits.webkit.org/297335@main