[EWS] Running layout tests in stress mode should skip tests marked as "Skip". https://bugs.webkit.org/show_bug.cgi?id=295827 rdar://155661448
Reviewed by Jonathan Bedard.

When a contributor posts a PR that modifies layout tests, EWS will run those
layout tests 100 times. This is usually a great way to see if a user broke
those tests with their change; however, consider the case of a contributor
landing a change with layout tests that are skipped. This could be because that
contributor has yet to implement the feature in full (e.g., they're landing PRs
as parts of a larger feature are developed). Those layout tests still run,
despite them being skipped, because they're passed in directly on the command
line for the `run-webkit-tests` invocation.

So, this modifies the behavior of `run-layout-tests-in-stress-mode` such that
skipped tests aren't run. We achieve this by passing in `--skipped always` in
that specific step.

* Tools/CISupport/ews-build/steps.py:
    (RunWebKitTests):
        -> (setLayoutTestCommand): Apply `--skipped always` based on whether we're running tests in stress mode or not.
* Tools/CISupport/ews-build/steps_unittest.py:
    (TestRunWebKitTestsInStressMode):
        -> (test_success): Modify unit test.
        -> (test_success_wk1): Ditto.
        -> (test_failure): Ditto.
        -> (test_success_additional_arguments): Ditto.
        -> (test_success): Ditto.
        -> (test_failure): Ditto.

Canonical link: https://commits.webkit.org/297361@main