Improve web platform test webvtt/rendering/cues-with-video/processing-model/enable_controls_reposition.html https://bugs.webkit.org/show_bug.cgi?id=295996 rdar://155897096
Reviewed by Jer Noble.

This patch fixes a few issues present in this test. First, I added the autoplay attribute to the expected
file to match the actual file.

In the expected file, I changed the cue box to be behind the video element, so that the transparent grey
background that WebKit places on top of the video when controls are showing shades the cue in the same way
in the expected and actual file.

I changed the source for the video from an mp4 to an mp3, and explicitly styled a border around the video
to ensure consistency between the files.

I added a call to takeScreenshot() in the expected file inside a seek event listener, to ensure the screenshot
is not taken until the test is complete. I did the same for the actual file, which previously called
takeScreenShotDelayed(1000). This should remove a bottleneck in testing speed.

With these changes, all browsers will still fail this test because the test assumes that the cue will move up
50px when the controls are showing, and each browser moves the cue a different number of pixels upward. I believe
50px is an arbitrary number that was chosen at an earlier time, and that the WebVTT spec does not specify an
offset. In the future, I think we should consider making a standard way for browsers to allow interacting with
the cue inside a test, so that we check that the cue is unobstructed by the controls in an nonarbitrary way.

* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/enable_controls_reposition-expected.html:
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/enable_controls_reposition-ref.html:
* LayoutTests/imported/w3c/web-platform-tests/webvtt/rendering/cues-with-video/processing-model/enable_controls_reposition.html:

Canonical link: https://commits.webkit.org/297425@main