AX: Stale value is served to assistive technologies for partially completed date inputs whose subfields change value https://bugs.webkit.org/show_bug.cgi?id=296011 rdar://155910722
Reviewed by Joshua Hoffman.

Date inputs are made up of sub-elements, which themselves are controls (specifically, spinbuttons). Prior to this commit,
we were not updating the cached date-value of date inputs that are partially completed (only some of the subfields have
a non-placeholder value). This is because the DOM code only notified accesibility of the input changing value when its
actual DOM value changed, and the DOM value of a date input is empty string until all fields are filled out.

With this commit, a hook is added into BaseDateAndTimeInputType to notify accessibility when a subfield changes.

* LayoutTests/accessibility/mac/stale-input-value-from-subfield-change-expected.txt: Added.
* LayoutTests/accessibility/mac/stale-input-value-from-subfield-change.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations: Disable new test.
* Source/WebCore/html/BaseDateAndTimeInputType.cpp:
(WebCore::BaseDateAndTimeInputType::didChangeValueFromControl):

Canonical link: https://commits.webkit.org/297435@main