[macOS 26] Using Writing tools in Quip fails when replacing text in a list after changing selection https://bugs.webkit.org/show_bug.cgi?id=296020 rdar://153077184
Reviewed by Abrar Rahman Protyasha.

In macOS 26, WritingToolsUI now uses `-[WKWebView readSelectionFromPasteboard:]` to simulate a paste
with rich text, when using Rewrite and Proofreading functionality in Safari, passing a unique
pasteboard to the web view.

However, this breaks Writing Tools when editing documents in Quip, which intercepts and prevents the
`paste` event in order to sanitize the DOM (and importantly, enforce a consistent DOM structure
necessary for their editor to save content properly). This is because, in this codepath that
triggers `Editor::readSelectionFromPasteboard`, we currently skip dispatching the `paste` event
altogether, and instead just read a document fragment from the pasteboard and insert it into the
DOM.

To fix this, align `readSelectionFromPasteboard` with other methods that are responsible for
triggering paste (`Editor::pasteAsQuotation`, `Editor::pasteAsPlainText`, `Editor::paste`) and call
`dispatchClipboardEvent` with an early return in the case where the page prevents default.

* Source/WebCore/editing/Editor.cpp:
(WebCore::Editor::dispatchClipboardEvent):
(WebCore::dispatchClipboardEvent): Deleted.

Turn this into a private static helper methods in `Editor`, so that we can call it from
`EditorCocoa.mm` below.

* Source/WebCore/editing/Editor.h:
* Source/WebCore/editing/cocoa/EditorCocoa.mm:
(WebCore::Editor::readSelectionFromPasteboard):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/PasteHTML.mm:
(TEST(PasteHTML, ReadSelectionFromPasteboardDispatchesPaste)):

Add an API test to exercise this change by using `-readSelectionFromPasteboard:` in a webpage that
intercepts `paste`, prevents default, and dumps the `DataTransfer` as plain text in the page.

Canonical link: https://commits.webkit.org/297439@main