[GTK][Tools] Forward the environment variables related to the GTK Scene Graph Kit (GSK) for the tests. https://bugs.webkit.org/show_bug.cgi?id=296023
Reviewed by Adrian Perez de Castro.

The GTK performance test bot has been failing the 'perf-test' step with the following error:
"Failed to realize renderer of type 'GskNglRenderer'for surface 'GdkWaylandToplevel': OpenGL ES 3.0 is not supported"

This happens because the GTK4 version on the Flatpak SDK defaults to the 'ngl' renderer, which requires OpenGL ES 3.0
But the hardware where this GTK perf bot runs (old Intel Ironlake) does not support OpenGL 3.0

To resolve this, we can explicitly set the GSK_RENDERER environment variable to 'gl' on the bot environment.
That tells GTK to use a GSK renderer compatible with OpenGL 2.1, which this hardware supports fine.

This patch whitelists any environment variable with a 'GSK_' prefix for the GTK port.

* Tools/Scripts/webkitpy/port/gtk.py:
(GtkPort.setup_environ_for_server):

Canonical link: https://commits.webkit.org/297454@main